--- src/components/organisms/Auth.tsx
@@ -1,143 +1,0 @@
-import { VFC } from "react";
-import { useHistory } from "react-router-dom";
-import { FaGithub, FaGoogle } from "react-icons/fa";
-import styled from "styled-components";
-import firebase from "firebase/app";
-
-import { COLOR } from "../../Themes/Color";
-import LoginImg from "../../images/logo.svg";
-import { DEVICE } from "../../Themes/Device";
-import { SignInButton } from "../atom/button/SignInButton";
-import { toastHandler } from "../../utils/toast";
-import { db, auth, githubProvider, googleProvider } from "../../libs/firebase";
-
-type Props = {
-  modalHandler: () => void;
-};
-
-const createUsersCollection = (user: firebase.User | null) => {
-  db.collection("users").add({
-    uid: user!.uid,
-    username: user!.displayName,
-    avatar: user!.photoURL,
-    comment: "",
-    likedPosts: [],
-    githubName: "",
-    twitterName: "",
-    blogUrl: "",
-  });
-};
-
-export const Auth: VFC<Props> = ({ modalHandler }) => {
-  const history = useHistory();
-
-  const signIn = (
-    provider:
-      | firebase.auth.GithubAuthProvider
-      | firebase.auth.GoogleAuthProvider
-  ) => {
-    auth
-      .signInWithPopup(provider)
-      .then(async ({ user }) => {
-        modalHandler();
-        // もしサインインしたユーザのidをもつコレクションが存在しなければusersコレクションに追加する
-        let userId = "";
-        const fetchUser = await db
-          .collection("users")
-          .where("uid", "==", `${user?.uid}`);
-        const res = await fetchUser.get();
-        res.forEach((doc) => {
-          if (doc.id) {
-            userId = doc.id;
-          }
-        });
-        if (!userId) {
-          createUsersCollection(user);
-        }
-        history.push(`/`);
-        toastHandler("success", "Sign In!!");
-      })
-      .catch((err) => {
-        console.log(err);
-      });
-  };
-
-  return (
-    <Modal>
-      <ModalMask onClick={modalHandler} />
-      <ModalInner>
-        <img src={LoginImg} alt="" width="240" />
-        <StyledModalTitle>レビューはきっと宝になる。</StyledModalTitle>
-        <ul>
-          <LoginItem>
-            <SignInButton onClick={() => signIn(githubProvider)}>
-              <StyledFaGithub />
-              Sign in with Github
-            </SignInButton>
-          </LoginItem>
-          <LoginItem>
-            <SignInButton onClick={() => signIn(googleProvider)}>
-              <StyledFaGoogle />
-              Sign in with Google
-            </SignInButton>
-          </LoginItem>
-        </ul>
-      </ModalInner>
-    </Modal>
-  );
-};
-
-const Modal = styled.div`
-  position: fixed;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-`;
-
-const ModalMask = styled.div`
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.2);
-`;
-
-const ModalInner = styled.div`
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  background-color: ${COLOR.WHITE};
-  width: 90vw;
-  max-width: 600px;
-  height: 400px;
-  border-radius: 10px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  box-shadow: 0 2.5rem 2rem -2rem hsl(200 50% 20% / 40%);
-
-  @media ${DEVICE.tabletL} {
-    width: 70vw;
-  }
-`;
-
-const StyledModalTitle = styled.h2`
-  font-size: 20px;
-  font-weight: bold;
-  color: ${COLOR.PRIMARY};
-  margin: 32px 0;
-`;
-
-const LoginItem = styled.li`
-  &:not(:last-child) {
-    padding-bottom: 20px;
-  }
-`;
-
-const StyledFaGithub = styled(FaGithub)`
-  margin-right: 10px;
-`;
-const StyledFaGoogle = styled(FaGoogle)`
-  margin-right: 10px;
-`;
