+++ src/components/organisms/CommentInputArea/CommentInputArea.tsx
@@ -1,11 +1,14 @@
 import { VFC, useState, memo } from "react";
-import styled from "styled-components";
-import { db } from "../../libs/firebase";
-import firebase from "firebase";
 
-import { DEVICE } from "../../Themes/Device";
-import { index as TextArea } from "../atom/textArea/index";
-import { PrimaryButton } from "../atom/button/PrimaryButton";
+import { db } from "../../../libs/firebase";
+import firebase from "firebase";
+import { PrimaryButton } from "../../atom/button/PrimaryButton";
+
+import {
+  StyledCommentInputArea,
+  StyledTextArea,
+  StyledButtonWrap,
+} from "./Styles";
 
 type Props = {
   uid: string;
@@ -48,31 +51,3 @@ export const CommentInputArea: VFC<Props> = memo(
     );
   }
 );
-
-const StyledCommentInputArea = styled.div`
-  text-align: right;
-`;
-
-const StyledTextArea = styled(TextArea)`
-  min-height: 240px;
-  width: calc(100% - 28px);
-  margin: 0 auto 10px;
-  border: none;
-  border-radius: 10px;
-  font-size: 16px;
-  line-height: 1.8;
-  padding: 20px 14px;
-
-  @media ${DEVICE.tabletL} {
-    width: 100%;
-    padding: 28px;
-  }
-`;
-
-const StyledButtonWrap = styled.div`
-  padding: 0 14px;
-
-  @media ${DEVICE.tabletL} {
-    padding: 0;
-  }
-`;
