+++ src/components/organisms/CommentOutputArea/CommentOutputArea.tsx
@@ -1,15 +1,23 @@
 import { VFC, memo } from "react";
-import styled from "styled-components";
-import { format } from "date-fns";
-import { RiDeleteBin6Line } from "react-icons/ri";
 
-import { index as Typography } from "../atom/typography/index";
-import { index as Title } from "../atom/title/index";
-import { IconWithName } from "../molecules/IconWithName";
-import { db } from "../../libs/firebase";
-import { DEVICE } from "../../Themes/Device";
-import { COLOR } from "../../Themes/Color";
-import { useCurrentUser } from "../../hooks/useCurrentUser";
+import { format } from "date-fns";
+
+import { index as Typography } from "../../atom/typography/index";
+import { index as Title } from "../../atom/title/index";
+import { IconWithName } from "../../molecules/IconWithName";
+import { db } from "../../../libs/firebase";
+
+import { useCurrentUser } from "../../../hooks/useCurrentUser";
+
+import {
+  StyledCommentWrap,
+  StyledCommentInner,
+  StyledComment,
+  StyledTimestamp,
+  StyledRiDeleteBin6Line,
+  StyledNameWithIcons,
+  StyledIcons,
+} from "./Styles";
 
 type Props = {
   postId: string;
@@ -68,68 +76,3 @@ export const CommentOutputArea: VFC<Props> = memo(({ postId, comments }) => {
     </StyledCommentWrap>
   );
 });
-
-const StyledCommentWrap = styled.div`
-  padding: 0 14px;
-
-  @media ${DEVICE.tabletL} {
-    padding: 0;
-  }
-`;
-
-const StyledCommentInner = styled.div`
-  background-color: ${COLOR.WHITE};
-  padding: 40px 20px;
-  box-sizing: border-box;
-  border-radius: 10px;
-  display: grid;
-  gap: 40px;
-
-  @media ${DEVICE.tabletL} {
-    padding: 40px;
-  }
-`;
-
-const StyledComment = styled.div`
-  display: grid;
-  gap: 10px;
-  padding-bottom: 32px;
-
-  &:not(:last-child) {
-    border-bottom: 1px solid ${COLOR.BACKGROUND};
-  }
-`;
-
-const StyledTimestamp = styled.span`
-  font-size: 12px;
-  color: ${COLOR.GRAY};
-`;
-
-const StyledRiDeleteBin6Line = styled(RiDeleteBin6Line)`
-  padding: 10px;
-  background-color: ${COLOR.BACKGROUND};
-  border-radius: 50%;
-  font-size: 12px;
-  transition: all 0.3s;
-
-  &:hover {
-    cursor: pointer;
-    background-color: ${COLOR.GRAY};
-    color: ${COLOR.WHITE};
-  }
-
-  @media ${DEVICE.tabletL} {
-    font-size: 16px;
-  }
-`;
-
-const StyledNameWithIcons = styled.div`
-  display: flex;
-  justify-content: space-between;
-  }
-`;
-
-const StyledIcons = styled.div`
-  display: flex;
-  gap: 10px;
-`;
