+++ src/components/organisms/Footer/Footer.tsx
@@ -1,12 +1,11 @@
 import { VFC } from "react";
-import styled from "styled-components";
 
-import { COLOR } from "../../Themes/Color";
-import { index as Link } from "../atom/link";
-import { index as Title } from "../atom/title/index";
-import { index as Copy } from "../atom/copywrite/index";
-import { LogoWithTypography } from "../molecules/LogoWithTypography";
-import { DEVICE } from "../../Themes/Device";
+import { index as Link } from "../../atom/link";
+import { index as Title } from "../../atom/title/index";
+import { index as Copy } from "../../atom/copywrite/index";
+import { LogoWithTypography } from "../../molecules/LogoWithTypography";
+
+import { StyledFooter, StyledFooterInner, StyledUl, StyledLi } from "./Styles";
 
 export const Footer: VFC = () => {
   return (
@@ -59,36 +58,3 @@ export const Footer: VFC = () => {
     </StyledFooter>
   );
 };
-
-const StyledFooter = styled.footer`
-  padding: 80px 0 40px;
-  border-top: 1px solid ${COLOR.BACKGROUND};
-
-  @media ${DEVICE.laptop} {
-    padding: 60px 0 20px;
-  }
-`;
-
-const StyledFooterInner = styled.div`
-  width: 90%;
-  margin: 0 auto;
-  display: grid;
-  grid-template-columns: 1fr;
-  gap: 40px 0;
-
-  @media ${DEVICE.tabletL} {
-    width: 70%;
-    max-width: 1200px;
-    grid-template-columns: 2fr 1fr 1fr;
-    gap: 0 40px;
-  }
-`;
-
-const StyledUl = styled.ul`
-  margin-top: 14px;
-`;
-
-const StyledLi = styled.li`
-  margin: 0.6rem 0;
-  font-size: 14px;
-`;
