+++ src/components/organisms/LikedPosts/LikedPosts.tsx
@@ -1,10 +1,10 @@
 import { VFC, useState, useEffect } from "react";
-import styled from "styled-components";
 
-import { index as LoadingIcon } from "../atom/loading/index";
-import { Article } from "../molecules/Article";
-import { db } from "../../libs/firebase";
-import { DEVICE } from "../../Themes/Device";
+import { index as LoadingIcon } from "../../atom/loading/index";
+import { Article } from "../../molecules/Article";
+import { db } from "../../../libs/firebase";
+
+import { StyledLikedPost } from "./Styles";
 
 type Props = {
   likedPosts: string[];
@@ -114,22 +114,3 @@ export const LikedPosts: VFC<Props> = ({ likedPosts }) => {
     </StyledLikedPost>
   );
 };
-
-const StyledLikedPost = styled.div`
-  width: 100%;
-  margin: 0 auto;
-  display: grid;
-  grid-template-columns: 1fr;
-  grid-gap: 16px 10px;
-
-  @media ${DEVICE.mobileM} {
-    grid-template-columns: 1fr 1fr;
-  }
-
-  @media ${DEVICE.laptop} {
-    width: 100%;
-    max-width: 1024px;
-    grid-template-columns: 1fr 1fr 1fr;
-    grid-gap: 10px 10px;
-  }
-`;
