+++ src/components/pages/CreatePost/Presenter.tsx
@@ -1,6 +1,6 @@
 import { VFC } from "react";
 import { Toaster } from "react-hot-toast";
-import { PostArea } from "../../organisms/PostArea";
+import { PostArea } from "../../organisms/PostArea/PostArea";
 
 export const Presenter: VFC = () => {
   return (
