+++ src/components/pages/EditPost/Presenter.tsx
@@ -1,7 +1,7 @@
 import { VFC } from "react";
 import { Toaster } from "react-hot-toast";
 
-import { PostArea } from "../../organisms/PostArea";
+import { PostArea } from "../../organisms/PostArea/PostArea";
 
 type Props = {
   editPostData: {
