+++ src/components/pages/Home/Presenter.tsx
@@ -4,7 +4,7 @@ import { Toaster } from "react-hot-toast";
 
 import { index as Title } from "../../atom/title/index";
 import { index as LoadingIcon } from "../../atom/loading/index";
-import { UserNameRegister } from "../../organisms/UserNameRegister";
+import { UserNameRegister } from "../../organisms/UserNameRegister/UserNameRegister";
 import { Article } from "../../molecules/Article";
 
 import {
