+++ src/components/pages/Profile/Presenter.tsx
@@ -4,9 +4,9 @@ import { Toaster } from "react-hot-toast";
 import { index as LoadingIcon } from "../../atom/loading/index";
 import { index as Icon } from "../../atom/icon/index";
 import { index as Title } from "../../atom/title/index";
-import { UserPost } from "../../organisms/UserPost";
-import { ProfileArea } from "../../organisms/ProfileArea";
-import { LikedPosts } from "../../organisms/LikedPosts";
+import { UserPost } from "../../organisms/UserPost/UserPost";
+import { ProfileArea } from "../../organisms/ProfileArea/ProfileArea";
+import { LikedPosts } from "../../organisms/LikedPosts/LikedPosts";
 import { Page404 } from "../../pages/Page404/Page404";
 
 import {
